+++ tests/testthat/test-pak.R
@@ -21,22 +21,43 @@ test_that("renv::install() works in projects using pak", {
 })
 
 test_that("renv::update() works in projects using pak", {
-  
+
   skip_on_cran()
   skip_on_windows()
   skip_if_not_installed("pak")
   pak <- renv_namespace_load("pak")
   renv_scope_options(renv.config.pak.enabled = TRUE)
   project <- renv_tests_scope()
-  
+
   # try installing an older version of 'breakfast'
   quietly(install("breakfast@0.1.0"))
   expect_true(renv_package_installed("breakfast"))
   expect_equal(renv_package_version("breakfast"), "0.1.0")
-  
+
   # try using 'update()' to update all installed packages
   quietly(update())
   expect_true(renv_package_installed("breakfast"))
   expect_equal(renv_package_version("breakfast"), "1.0.0")
-  
+
+})
+
+# https://github.com/rstudio/renv/issues/2282
+test_that("renv::init() routes through pak when pak.enabled = TRUE", {
+
+  renv_tests_scope("bread")
+  renv_scope_options(renv.config.pak.enabled = TRUE)
+
+  pak_install_packages <- NULL
+  local_mocked_bindings(
+    renv_pak_init = function(...) invisible(NULL),
+    renv_pak_install = function(packages, ...) {
+      pak_install_packages <<- packages
+      invisible(NULL)
+    }
+  )
+
+  quietly(init())
+
+  expect_true("bread" %in% pak_install_packages)
+
 })